+++ lib/index.ts
@@ -45,6 +45,7 @@ export {
   resolveModulePattern,
 } from './modules/module-loader';
 export type { LoadedModule } from './modules/module-loader';
+export { collectModuleEntities } from './modules/collect-entities';
 export type {
   ModuleState,
   ModuleInstall,
@@ -71,4 +72,5 @@ export * from './utilities/errors';
 export * from './utilities/config-service';
 export * from './interfaces/type-error';
 export * from './interfaces/utils';
+export type { LitebOptions } from './core/liteb';
 export { Liteb };